smart-farm_backend-ms2/requirement.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
gunicorn==20.1.0
pillow==10.1.0
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas
scikit-learn==1.3.2
joblib==1.3.2
tensorflow
opencv-python-headless
google-generativeai==0.3.2
werkzeug==2.3.7
pytest==7.4.3
pytest-flask==1.3.0